refactor(activity): extract formatActivity to lib and fix e2e tests (#1085)
* test: add unit tests for formatActivity event formatter in metrics/activity route

* fix(e2e): use case-insensitive heading match for DASHBOARD h1

The DashboardHeader renders the h1 as 'DASHBOARD' (all-caps via
design), but the Playwright selectors used an exact-case match
{ name: "Dashboard" } which never matched.

Changed all three heading assertions to /dashboard/i regex so they
match the actual DOM output and unblock every failing E2E test.

* fix(e2e): fix all Playwright test failures across 3 spec files

landing.spec.js:
- Remove assertion for 'DevTrack' h1 (hero reads 'YOUR CODE HAS A PULSE')
- Add .first() to 'Sign in with GitHub' link selectors (2 links on page,
  strict mode violation)
- Replace 'View on GitHub' link with '★ Star on GitHub' (actual text)

auth-bypass.spec.js:
- Add .first() to 'Sign in with GitHub' link (strict mode violation)
- Change heading assertions from exact 'Dashboard' to /dashboard/i regex

dashboard-widgets.spec.js:
- Add mock for /api/goals/sync so GoalTracker doesn't hang waiting for
  a real Supabase call (Goals heading was never rendering in CI)
- Add mocks for additional API routes hit on dashboard load

* refactor(activity): extract formatActivity to lib module to fix build

Next.js App Router only permits HTTP-verb exports (GET, POST, etc.)
from route files. Exporting 'formatActivity' directly from route.ts
caused the CI build to fail:

  Route "src/app/api/metrics/activity/route.ts" does not match the
  required types of a Next.js Route.
  "formatActivity" is not a valid Route export field.

Fix:
- Move formatActivity, its types (ActivityItem, RawEvent, ActivityType),
  and SUPPORTED_EVENT_TYPES into src/lib/activity-formatter.ts
- Import from the new module in route.ts (no functional change)
- Update test/formatActivity.test.ts to import from the lib module
  (removes now-unnecessary vi.mock calls for supabase/auth/next-auth)

---------
